Render cards with different adapters on a page.
This commit allows us to specify an adapter on a per card basis, so that
a Container can have multiple cards that are inline and/or iframe-d. It
makes more sense for each card to specify its own
adapter, thus the Container should always be inline so that it can
render into the right slot.