Dynamically add a new card to a container.
First we had to define a `CardManagerService`, which initialized with a
port and had an `addCard` method defined. This method was then available
to the container instance inside of its resolved promise (using
`waitForLoad()`)... as
`card.sandbox.capabilities.cardManager.addCard()`. This passes the
option into the container card reference itself, where we setup the slot
and append the new card into the DOM.